# Doing the aggregation by asset in numpy, not in Python