This is a lot faster. For instance on gemsun02 the runtime of our demo went down from 395 s to 208 s and  the memory allocation from 30 MB to 22 MB. In the case of Turkey with 100 years I saw a change from  81,373 s -> 47,264 s.

The demos are running here: https://ci.openquake.org/job/zdevel_oq-risklib/1400/