@ptormene 2016-03-04
Apart from a minor note about naming conventions, LGTM